pyproject.toml
[tool.poetry]
name = "todo-app-graphql"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
strawberry-graphql = "^0.216.1"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
pydantic = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
